### 2024-10-18 #7367 @editorialbot
[PRE REVIEW]: Sigma: Uncertainty Propagation for C++
**Submitting author:** @jwaldrop107 (<a href="http://orcid.org/0000-0003-0442-193X">Jonathan Waldrop</a>)
**Repository:** https://github.com/QCUncertainty/sigma
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.1
