### 2024-11-11 #7404 @baxmittens
## Documentation
`A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?`

There is a motivation section but it is a bit short. But a statement of need is given in the paper. This could be used as a template for the documentation.

`Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support`

I think community guidelines are missing.
