### 2024-11-20 189aabb
expand motivation and rename as statement of need
