+++ search/all_6.js
@@ -1,5 +1,20 @@
 var searchData=
 [
-  ['hyperbolic_2ehpp_0',['hyperbolic.hpp',['../hyperbolic_8hpp.html',1,'']]],
-  ['hypot_1',['hypot',['../namespacesigma.html#a2ca53b359261368fe6bad807f92893fb',1,'sigma::hypot(const Uncertain&lt; T &gt; &amp;a, const Uncertain&lt; T &gt; &amp;b)'],['../namespacesigma.html#a799d4d86ba1458a66bd391c685e1f993',1,'sigma::hypot(const Uncertain&lt; T &gt; &amp;a, const U &amp;b)'],['../namespacesigma.html#a181b0d75d0f1bda0bf7be58cb2236032',1,'sigma::hypot(const U &amp;a, const Uncertain&lt; T &gt; &amp;b)']]]
+  ['cbrt_0',['cbrt',['../namespacesigma.html#ac79de8ee1df5ced403936214add4e64e',1,'sigma']]],
+  ['ceil_1',['ceil',['../namespacesigma.html#aae07ce73db5514f76faadd45497679fd',1,'sigma']]],
+  ['code_20of_20conduct_2',['Code of Conduct',['../md_docs_2contributing.html#code-of-conduct',1,'Code of Conduct'],['../md_docs_2code__of__conduct.html',1,'Contributor Covenant Code of Conduct']]],
+  ['comparison_3',['Equality and Comparison',['../md_docs_2examples.html#equality-and-comparison',1,'']]],
+  ['complex_2ehpp_4',['complex.hpp',['../complex_8hpp.html',1,'']]],
+  ['conduct_5',['Conduct',['../md_docs_2contributing.html#code-of-conduct',1,'Code of Conduct'],['../md_docs_2code__of__conduct.html',1,'Contributor Covenant Code of Conduct']]],
+  ['conj_6',['conj',['../namespacesigma.html#a0a033b5557b1f726f89fa4cba5140e92',1,'sigma']]],
+  ['construction_20of_20uncertain_20variables_7',['Construction of Uncertain Variables',['../md_docs_2examples.html#construction-of-uncertain-variables',1,'']]],
+  ['contribute_8',['Ways to Contribute',['../md_docs_2contributing.html#ways-to-contribute',1,'']]],
+  ['contributing_9',['Contributing',['../index.html#contributing',1,'']]],
+  ['contributor_20covenant_20code_20of_20conduct_10',['Contributor Covenant Code of Conduct',['../md_docs_2code__of__conduct.html',1,'']]],
+  ['contributor_20guidelines_11',['Contributor Guidelines',['../md_docs_2contributing.html',1,'']]],
+  ['copysign_12',['copysign',['../namespacesigma.html#acd31fdaf16ebab84607b4f004204a556',1,'sigma::copysign(const Uncertain&lt; T &gt; &amp;a, const Uncertain&lt; T &gt; &amp;b)'],['../namespacesigma.html#aab0b028f2d1d0c35269b0d0c9f75ccfa',1,'sigma::copysign(const Uncertain&lt; T &gt; &amp;a, const U &amp;b)'],['../namespacesigma.html#a60742125f0d5c889ac848e306876404f',1,'sigma::copysign(const U &amp;a, const Uncertain&lt; T &gt; &amp;b)']]],
+  ['correction_13',['1. Correction',['../md_docs_2code__of__conduct.html#autotoc_md1-correction',1,'']]],
+  ['cos_14',['cos',['../namespacesigma.html#a8d0718197bd681b48cfbc005d6629920',1,'sigma']]],
[…]
