+++ search/pages_5.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['sigma_0',['Sigma',['../index.html',1,'']]]
+];
