+++ search/all_11.js
@@ -1,5 +1,8 @@
 var searchData=
 [
-  ['value_5ft_0',['value_t',['../classsigma_1_1detail___1_1Setter.html#a62c1d0537a6ee848492a90d64782c57a',1,'sigma::detail_::Setter::value_t'],['../classsigma_1_1Uncertain.html#ab9d55c3ae3c72960a33db279c96cb952',1,'sigma::Uncertain::value_t']]],
-  ['variables_1',['Construction of Uncertain Variables',['../md_docs_2examples.html#construction-of-uncertain-variables',1,'']]]
+  ['permanent_20ban_0',['4. Permanent Ban',['../md_docs_2code__of__conduct.html#autotoc_md4-permanent-ban',1,'']]],
+  ['pi_1',['pi',['../namespacesigma_1_1detail__.html#a8d8ed28e719cef3b02f64c0d74a773ac',1,'sigma::detail_']]],
+  ['pledge_2',['Our Pledge',['../md_docs_2code__of__conduct.html#our-pledge',1,'']]],
+  ['pow_3',['pow',['../namespacesigma.html#ac240ea50be68f2eb2ee4b6abf0d063c7',1,'sigma::pow(const Uncertain&lt; T &gt; &amp;a, const U &amp;exp)'],['../namespacesigma.html#a0b3162c4a58cc30e72f7c4ddf3dfc1f6',1,'sigma::pow(const Uncertain&lt; T &gt; &amp;a, const Uncertain&lt; T &gt; &amp;exp)']]],
+  ['propagation_4',['Linear Error Propagation',['../md_docs_2background.html#linear-error-propagation',1,'']]]
 ];
