+++ search/all_b.js
@@ -1,22 +1,5 @@
 var searchData=
 [
-  ['of_20uncertain_20variables_0',['Construction of Uncertain Variables',['../md_docs_2examples.html#construction-of-uncertain-variables',1,'']]],
-  ['operation_5fcommon_2ehpp_1',['operation_common.hpp',['../operation__common_8hpp.html',1,'']]],
-  ['operations_2',['Mathematical Operations',['../md_docs_2examples.html#mathematical-operations',1,'']]],
-  ['operations_2ehpp_3',['operations.hpp',['../operations_8hpp.html',1,'']]],
-  ['operator_21_3d_4',['operator!=',['../classsigma_1_1Uncertain.html#ab838d18dba696b69de1af74212f04a85',1,'sigma::Uncertain::operator!=()'],['../classsigma_1_1Uncertain.html#ab838d18dba696b69de1af74212f04a85',1,'sigma::operator!=(const Uncertain&lt; ValueType1 &gt; &amp;lhs, const Uncertain&lt; ValueType2 &gt; &amp;rhs)']]],
-  ['operator_2a_5',['operator*',['../namespacesigma.html#aad6843bd1c16f459b4d09ca515bc677d',1,'sigma::operator*(const Uncertain&lt; T &gt; &amp;lhs, const Uncertain&lt; T &gt; &amp;rhs)'],['../namespacesigma.html#ab15114e4948864693c0384f79a23897a',1,'sigma::operator*(const Uncertain&lt; T &gt; &amp;lhs, double rhs)'],['../namespacesigma.html#a2197d22cf204728006d82f03962a8191',1,'sigma::operator*(double lhs, const Uncertain&lt; T &gt; &amp;rhs)']]],
-  ['operator_2a_3d_6',['operator*=',['../namespacesigma.html#a920dd90b639d9e3737931d2d2af297cd',1,'sigma::operator*=(Uncertain&lt; T &gt; &amp;lhs, const Uncertain&lt; T &gt; &amp;rhs)'],['../namespacesigma.html#a9418df48959cb9f98e1f61c7f7ade893',1,'sigma::operator*=(Uncertain&lt; T &gt; &amp;lhs, double rhs)']]],
-  ['operator_2b_7',['operator+',['../namespacesigma.html#a987e777dfd93f18538ac4a9cb997c110',1,'sigma::operator+(const Uncertain&lt; T &gt; &amp;lhs, const Uncertain&lt; T &gt; &amp;rhs)'],['../namespacesigma.html#a788871296627e9850ba6abc900d18303',1,'sigma::operator+(const Uncertain&lt; T &gt; &amp;lhs, double rhs)'],['../namespacesigma.html#ad57001da4c4ffb072f578e7ee2df757f',1,'sigma::operator+(double lhs, const Uncertain&lt; T &gt; &amp;rhs)']]],
-  ['operator_2b_3d_8',['operator+=',['../namespacesigma.html#aff88b15205b29f22c6fd500a51ac5c50',1,'sigma::operator+=(Uncertain&lt; T &gt; &amp;lhs, const Uncertain&lt; T &gt; &amp;rhs)'],['../namespacesigma.html#a3f750bcb59fa457e6d54b53391823aeb',1,'sigma::operator+=(Uncertain&lt; T &gt; &amp;lhs, double rhs)']]],
[…]
