### 2024-11-20 #7404 @jwaldrop107
Hi all,

Sorry about the delayed response. I've pushed some changes based on the comments made so far (wording in the paper, statement of need in the documentation, and community guidelines).

Thanks for taking the time to review this, and I'm looking forward to any additional comments/recommendations.
