### 2024-11-27 #7404 @baxmittens
@jwaldrop107
I created a pull request for the paper.md, where I fixed 2-3 typos. Be very careful accepting these since I am not a native speaker. The last one (operated for/by) could be wrong.
