### 2024-12-02 4e399f3
Update paper/paper.md
