+++ paper/paper.md
@@ -190,7 +190,7 @@ Additional functionality is possible, though not currently ensured.
 # Acknowledgements
 
 This work was supported by the Ames National Laboratory's Laboratory Directed 
-Research and Development (LDRD) program. The Ames Laboratory is operated by the
+Research and Development (LDRD) program. The Ames Laboratory is operated for the
 U.S. DOE by Iowa State University under contract # DE-AC02-07CH11358.
 
 # References
