### 2024-12-04 #7404 @vissarion
@YehorYudinIPP what is the progress of your review? There are some some points in your checklists that are not checked out. Are there any issues?
