### 2024-12-06 #7404 @YehorYudinIPP
> @YehorYudinIPP what is the progress of your review? There are some some points in your checklists that are not checked out. Are there any issues?

I reviewed everything written and only have minor comments which I will compile and share later.
However, I still haven't installed and reproduced tests etc. which I can do in the end of the next week the earliest.
