### 2024-12-19 #7404 @YehorYudinIPP
I am currently installing and running library usage examples on a fresh Ubuntu 24.04 under Windows Susbsytems from Linux, and, due to some obvious omission of some preliminary steps for the library installation description, I encounter issues with doxygen run and documentation generation.
It is reasonable not to include automatic support for doxygen installation in the library, still, to be sure that everything installation is reproducible I nevertheless want to figure out what is wrong in my set-up.
Despite installing doxygen and exporting the required environmental variable, I encounter `Could NOT find Doxygen (missing: DOXYGEN_EXECUTABLE)` error
It seemed it could be due to a number of reasons and I haven't resolved the issue yet.
