### 2025-01-16 9d1edf2
Update Eigen, tests should pass
