+++ basic_8hpp.js
@@ -1,16 +1,16 @@
 var basic_8hpp =
 [
-    [ "abs", "basic_8hpp.html#add954174f373873d59d093ceb3e82af5", null ],
-    [ "abs2", "basic_8hpp.html#aaa0794dcf432114a9467603ac7748c7c", null ],
-    [ "ceil", "basic_8hpp.html#aae07ce73db5514f76faadd45497679fd", null ],
-    [ "copysign", "basic_8hpp.html#a60742125f0d5c889ac848e306876404f", null ],
-    [ "copysign", "basic_8hpp.html#aab0b028f2d1d0c35269b0d0c9f75ccfa", null ],
-    [ "copysign", "basic_8hpp.html#acd31fdaf16ebab84607b4f004204a556", null ],
-    [ "fabs", "basic_8hpp.html#ac4d519de22722a51cda0bc3c92960e89", null ],
-    [ "floor", "basic_8hpp.html#aa33c27c2144a61b5061e6a4902349123", null ],
-    [ "fmod", "basic_8hpp.html#aff47e6b43a710dc89637c5e6ae1e1c60", null ],
-    [ "fmod", "basic_8hpp.html#a09be7785a22e02bd9b1ff1e44f3b5072", null ],
-    [ "fmod", "basic_8hpp.html#a878dc31b111a5c5db62a8985188b8247", null ],
-    [ "round", "basic_8hpp.html#a940424e34df1bad854b05c139ce73648", null ],
-    [ "trunc", "basic_8hpp.html#ae431d0534d90713daf06f763dd523277", null ]
+    [ "sigma::abs", "namespacesigma.html#add954174f373873d59d093ceb3e82af5", null ],
+    [ "sigma::abs2", "namespacesigma.html#aaa0794dcf432114a9467603ac7748c7c", null ],
+    [ "sigma::ceil", "namespacesigma.html#aae07ce73db5514f76faadd45497679fd", null ],
+    [ "sigma::copysign", "namespacesigma.html#a60742125f0d5c889ac848e306876404f", null ],
+    [ "sigma::copysign", "namespacesigma.html#aab0b028f2d1d0c35269b0d0c9f75ccfa", null ],
+    [ "sigma::copysign", "namespacesigma.html#acd31fdaf16ebab84607b4f004204a556", null ],
+    [ "sigma::fabs", "namespacesigma.html#ac4d519de22722a51cda0bc3c92960e89", null ],
+    [ "sigma::floor", "namespacesigma.html#aa33c27c2144a61b5061e6a4902349123", null ],
+    [ "sigma::fmod", "namespacesigma.html#aff47e6b43a710dc89637c5e6ae1e1c60", null ],
+    [ "sigma::fmod", "namespacesigma.html#a09be7785a22e02bd9b1ff1e44f3b5072", null ],
+    [ "sigma::fmod", "namespacesigma.html#a878dc31b111a5c5db62a8985188b8247", null ],
+    [ "sigma::round", "namespacesigma.html#a940424e34df1bad854b05c139ce73648", null ],
+    [ "sigma::trunc", "namespacesigma.html#ae431d0534d90713daf06f763dd523277", null ]
 ];
