+++ hyperbolic_8hpp.js
@@ -1,9 +1,9 @@
 var hyperbolic_8hpp =
 [
-    [ "acosh", "hyperbolic_8hpp.html#ae0468bc7af0c3d24c8ecc0c38d8f2ef6", null ],
-    [ "asinh", "hyperbolic_8hpp.html#aad587ce36b20a927d652a01148dc343c", null ],
-    [ "atanh", "hyperbolic_8hpp.html#ad13293749576a5d73b834c598e2936c6", null ],
-    [ "cosh", "hyperbolic_8hpp.html#a8f8ded25467d58703d7ec390526bb2bb", null ],
-    [ "sinh", "hyperbolic_8hpp.html#a2e4a33ee5b89ddb9a1f023d6341d75a3", null ],
-    [ "tanh", "hyperbolic_8hpp.html#aa5a3cb36953c116658b16ddc790db42d", null ]
+    [ "sigma::acosh", "namespacesigma.html#ae0468bc7af0c3d24c8ecc0c38d8f2ef6", null ],
+    [ "sigma::asinh", "namespacesigma.html#aad587ce36b20a927d652a01148dc343c", null ],
+    [ "sigma::atanh", "namespacesigma.html#ad13293749576a5d73b834c598e2936c6", null ],
+    [ "sigma::cosh", "namespacesigma.html#a8f8ded25467d58703d7ec390526bb2bb", null ],
+    [ "sigma::sinh", "namespacesigma.html#a2e4a33ee5b89ddb9a1f023d6341d75a3", null ],
+    [ "sigma::tanh", "namespacesigma.html#aa5a3cb36953c116658b16ddc790db42d", null ]
 ];
