+++ search/all_7.js
@@ -1,9 +1,9 @@
 var searchData=
 [
   ['degrees_0',['degrees',['../namespacesigma.html#ac8aa7a731dc93ac48f683999b3f712e8',1,'sigma']]],
-  ['dep_5fsd_5fptr_1',['dep_sd_ptr',['../classsigma_1_1detail___1_1Setter.html#a7b26186a2757c8eff7a15134a26e4c8e',1,'sigma::detail_::Setter::dep_sd_ptr'],['../classsigma_1_1Uncertain.html#acc40534d87fdba24215242f6021babe0',1,'sigma::Uncertain::dep_sd_ptr']]],
-  ['dep_5fsd_5ft_2',['dep_sd_t',['../classsigma_1_1detail___1_1Setter.html#a51b1eb8dc2255f1fa71b64f9afc13006',1,'sigma::detail_::Setter::dep_sd_t'],['../classsigma_1_1Uncertain.html#a603fe75d1c25677f33be269015a61639',1,'sigma::Uncertain::dep_sd_t']]],
+  ['dep_5fsd_5fptr_1',['dep_sd_ptr',['../classsigma_1_1detail___1_1Setter.html#a317752307a6665ce2ab3c577755599e7',1,'sigma::detail_::Setter::dep_sd_ptr'],['../classsigma_1_1Uncertain.html#acc40534d87fdba24215242f6021babe0',1,'sigma::Uncertain::dep_sd_ptr']]],
+  ['dep_5fsd_5ft_2',['dep_sd_t',['../classsigma_1_1detail___1_1Setter.html#acd92e5a8d988dd83e97926c58f4d5ebb',1,'sigma::detail_::Setter::dep_sd_t'],['../classsigma_1_1Uncertain.html#a603fe75d1c25677f33be269015a61639',1,'sigma::Uncertain::dep_sd_t']]],
   ['deps_3',['deps',['../classsigma_1_1Uncertain.html#a8f8c4a9313c5cb54e1eb0949a5214f13',1,'sigma::Uncertain']]],
-  ['deps_5fmap_5ft_4',['deps_map_t',['../classsigma_1_1detail___1_1Setter.html#a88a65cd262fe0918a1b290e4da8e3506',1,'sigma::detail_::Setter::deps_map_t'],['../classsigma_1_1Uncertain.html#a52e06c981ac599c8fccdd31e03863f0e',1,'sigma::Uncertain::deps_map_t']]],
+  ['deps_5fmap_5ft_4',['deps_map_t',['../classsigma_1_1detail___1_1Setter.html#ad786977eb876807acd9fb35bae3f84fc',1,'sigma::detail_::Setter::deps_map_t'],['../classsigma_1_1Uncertain.html#a52e06c981ac599c8fccdd31e03863f0e',1,'sigma::Uncertain::deps_map_t']]],
   ['detail_5f_3a_3asetter_5',['Setter',['../classsigma_1_1Uncertain.html#a7591e2d7ef0c929d5e9b38697c7c8f4f',1,'sigma::Uncertain']]]
 ];
