### 2025-01-28 3f7aa8b
clean up examples in paper
