### 2025-01-29 a05bfc0
Add comparison to other packages
