### 2025-01-31 5a298e4
Testing tensor contractions
