+++ tests/unit_tests/test_eigen.cpp
@@ -142,50 +142,66 @@ TEMPLATE_TEST_CASE("Eigen Matrix with Uncertain Elements", "", sigma::UFloat,
         using matrix_t = Eigen::Tensor<testing_t, 2>;
         using index_t  = Eigen::IndexPair<int>;
 
-        // Replicate TW Tests
+        // Various Input Values
         scalar_t scalar{};
-        scalar(0) = 42.0;
+        scalar(0) = u(42.0);
 
         vector_t vector(2);
-        vector(0) = 0.0;
-        vector(1) = 1.0;
+        vector(0) = u(0.0);
+        vector(1) = u(1.0);
 
         matrix_t matrix(2, 2);
-        matrix(0, 0) = 1.0;
-        matrix(0, 1) = 2.0;
-        matrix(1, 0) = 3.0;
-        matrix(1, 1) = 4.0;
+        matrix(0, 0) = u(1.0);
+        matrix(0, 1) = u(2.0);
+        matrix(1, 0) = u(3.0);
+        matrix(1, 1) = u(4.0);
 
-        std::cout << "Inputs" << std::endl;
-        std::cout << scalar << "\n\n";
-        std::cout << vector << "\n\n";
-        std::cout << matrix << "\n\n";
+        SECTION("i,i->") {
+            scalar_t corr{};
+            corr(0) = vector(0) * vector(0) + vector(1) * vector(1);
 
-        // i,i->
-        std::cout << "i,i->" << std::endl;
-        Eigen::array<index_t, 1> m0 = {index_t(0, 0)};
-        scalar_t scalar_rv1         = vector.contract(vector, m0);
-        std::cout << scalar_rv1 << "\n\n";
+            Eigen::array<index_t, 1> m{index_t(0, 0)};
+            scalar_t rv = vector.contract(vector, m);
+            REQUIRE(rv(0) == corr(0));
+        }
+        SECTION("ij,ij->") {
+            scalar_t corr{};
+            corr(0) = matrix(0, 0) * matrix(0, 0) +
+                      matrix(0, 1) * matrix(0, 1) +
+                      matrix(1, 0) * matrix(1, 0) + matrix(1, 1) * matrix(1, 1);
 
-        // ij,ij->
-        std::cout << "ij,ij->" << std::endl;
-        Eigen::array<index_t, 2> m1 = {index_t(0, 0), index_t(1, 1)};
-        scalar_t scalar_rv2         = matrix.contract(matrix, m1);
-        std::cout << scalar_rv2 << "\n\n";
+            Eigen::array<index_t, 2> m{index_t(0, 0), index_t(1, 1)};
+            scalar_t rv = matrix.contract(matrix, m);
+            REQUIRE(rv(0) == corr(0));
+        }
+        SECTION("ki,kj->ij") {
+            matrix_t corr(2, 2);
+            corr(0, 0) =
+              matrix(0, 0) * matrix(0, 0) + matrix(1, 0) * matrix(1, 0);
+            corr(0, 1) =
+              matrix(0, 0) * matrix(0, 1) + matrix(1, 0) * matrix(1, 1);
+            corr(1, 0) =
+              matrix(0, 1) * matrix(0, 0) + matrix(1, 1) * matrix(1, 0);
+            corr(1, 1) =
+              matrix(0, 1) * matrix(0, 1) + matrix(1, 1) * matrix(1, 1);
 
-        // ki,kj->ij
-        std::cout << "ki,kj->ij" << std::endl;
-        Eigen::array<index_t, 1> m2 = {index_t(0, 0)};
-        matrix_t matrix_rv1         = matrix.contract(matrix, m2);
-        std::cout << matrix_rv1 << "\n\n";
[…]
