### 2025-02-10 #7404 @YehorYudinIPP
Hi @vissarion! Yes, I recommend acceptance
