### 2025-02-11 #7404 @jwaldrop107
The tag is v1.0 and the release is [here](https://github.com/QCUncertainty/sigma/releases/tag/v1.0). The Zenodo DOI is [10.5281/zenodo.14852780](https://doi.org/10.5281/zenodo.14852780).
