### 2025-02-13 #7404 @jwaldrop107
@vissarion, the title and affiliations should be correct now.
