### 2025-02-13 #7404 @crvernon
:wave: @jwaldrop107 - well done.  I just need you to fix the following before I move to accept this submission for publication:

- [ ] LINE 91:  the "p" in "python" should be capitalized.  You can do this in your bib file by placing the characters you wish to maintain formatting of in curly brackets.

No new release is required.  Let me know when this has been fixed. Thanks.
