+++ paper/paper.bib
@@ -15,7 +15,7 @@
 
 @misc{uncertainties,
   author = {Eric O. Lebigot},
-  title = {Uncertainties: a Python package for calculations with uncertainties},
+  title = {Uncertainties: a {P}ython package for calculations with uncertainties},
   year = {2009},
   publisher = {GitHub},
   journal = {GitHub repository},
