### 2025-02-13 #7404 @jwaldrop107
@crvernon The capitalization has been fixed with [6d9c1d7](https://github.com/QCUncertainty/sigma/pull/13/commits/6d9c1d7cb9da11d87b1aa88d7660bfb3bad03d0e).
